review fixes #5
