+++ src/main/java/ru/hh/resumebuilderbot/message/handler/AnswerMessageHandler.java
@@ -3,11 +3,11 @@ package ru.hh.resumebuilderbot.message.handler;
 import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.CurrentUserState;
 import ru.hh.resumebuilderbot.QuestionsStorage;
-import ru.hh.resumebuilderbot.UserDataStorage;
 import ru.hh.resumebuilderbot.question.generator.FixedQuestionGenerator;
 import ru.hh.resumebuilderbot.question.generator.QuestionGenerator;
 import ru.hh.resumebuilderbot.question.generator.QuestionGeneratorByNumber;
 import ru.hh.resumebuilderbot.texts.storage.TextId;
+import ru.hh.resumebuilderbot.user.data.storage.UserDataStorage;
 
 import java.util.Queue;
