+++ src/main/java/ru/hh/resumebuilderbot/message/handler/StartMessageHandler.java
@@ -2,11 +2,11 @@ package ru.hh.resumebuilderbot.message.handler;
 
 import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.ChatId;
-import ru.hh.resumebuilderbot.UserDataStorage;
 import ru.hh.resumebuilderbot.question.generator.FirstQuestionGenerator;
 import ru.hh.resumebuilderbot.question.generator.FixedQuestionGenerator;
 import ru.hh.resumebuilderbot.question.generator.QuestionGenerator;
 import ru.hh.resumebuilderbot.texts.storage.TextId;
+import ru.hh.resumebuilderbot.user.data.storage.UserDataStorage;
 
 import java.util.Queue;
